requirements.txt
# =========================
# Core Web
# =========================
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==21.2.0

# =========================
# Database / ORM
# =========================
SQLAlchemy==2.0.32
alembic==1.13.2
psycopg[binary]==3.3.2

# =========================
# Auth / Security
# =========================
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
itsdangerous==2.2.0

# =========================
# Validation / Settings
# =========================
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1

# =========================
# Templates / Forms
# =========================
jinja2==3.1.4
python-multipart==0.0.9

# =========================
# Utils
# =========================
email-validator==2.2.0
pillow==10.4.0

# =========================
# PDF + IA
# =========================
pypdf==5.1.0
openai==1.57.4
httpx>=0.27.0

# =========================
# Relatórios / Exportações
# =========================
openpyxl==3.1.5
reportlab==4.1.0

